==> demod/toovood/riigi_teataja_sisuotsing/legacy/02_download_xml_sources.ipynb <==
import time

from random import uniform
from pandas import read_csv
from tqdm.auto import tqdm
from urllib.request import urlretrieve

SLEEP_MIN = 0.5
SLEEP_MAX = 2.5

## I. Crawl all state laws

sources = read_csv('results/state_laws.csv', header=0)
TARGET_PATH = 'results/xml_sources/state_laws'

for idx, row in tqdm(sources.iterrows(), total=len(sources)):
    print(row['xml_source'])
    output_file = f"{TARGET_PATH}/{row['xml_source'].split('/')[-1]}"
    print(output_file)
    _, status = urlretrieve(row['xml_source'], output_file) 
    time.sleep(uniform(SLEEP_MIN, SLEEP_MAX))

## II. Crawl all government regulations

sources = read_csv('results/government_regulations.csv', header=0)
TARGET_PATH = 'results/xml_sources/government_regulations'

for idx, row in tqdm(sources.iterrows(), total=len(sources)):
    print(row['xml_source'])
    output_file = f"{TARGET_PATH}/{row['xml_source'].split('/')[-1]}"
    print(output_file)
    _, status = urlretrieve(row['xml_source'], output_file) 
    time.sleep(uniform(SLEEP_MIN, SLEEP_MAX))

## III. Crawl all local government acts

sources = read_csv('results/local_government_acts.csv', header=0)
TARGET_PATH = 'results/xml_sources/local_government_acts'

for idx, row in tqdm(sources.iterrows(), total=len(sources)):
    print(row['xml_source'])
    output_file = f"{TARGET_PATH}/{row['xml_source'].split('/')[-1]}"
    print(output_file)
    _, status = urlretrieve(row['xml_source'], output_file) 
    time.sleep(uniform(SLEEP_MIN, SLEEP_MAX))

## IV. Crawl all government orders

sources = read_csv('results/government_orders.csv', header=0)
TARGET_PATH = 'results/xml_sources/government_orders'

for idx, row in tqdm(sources.iterrows(), total=len(sources)):
    print(row['xml_source'])
    output_file = f"{TARGET_PATH}/{row['xml_source'].split('/')[-1]}"
    print(output_file)
    _, status = urlretrieve(row['xml_source'], output_file) 
    time.sleep(uniform(SLEEP_MIN, SLEEP_MAX))
