==> demod/toovood/riigi_teataja_sisuotsing/legacy/03_exploration_of_xml_documents.ipynb <==
## Observations about the document structure

Meil on olemas kolme liiki dokumente
* tavadokumendid
* tühjad dokumendid milles on XML viide originaalile
* seaduste ja lepingute kaaskirjad, millel on HTML viide originaalile

Tavadokumendid võivad olla jagatud alamosadeks mis on nummerdatud, aga need alamosad ise ei sisalda teksti.
* Dokumendi tekstiline osa on alati XML-elemendi `<sisuText>` sees.
* XML-element `<sisuTekst>` sisaldab alamelemente `<tavatekst>`, `<HTMLKonteiner>`, `<viide>`
