==> demod/toovood/riigi_teataja_sisuotsing/legacy/04_crawl_additional_references.ipynb <==
import os
import re
import glob

from lxml import etree
from pandas import concat
from tqdm.auto import tqdm

from xml_ops import remove_namespaces
from xml_ops import extract_references_from_document

## I. Extract all references from XML documents

results = []

INPUT_DIR = 'results/xml_sources/state_laws/'
for file in tqdm(glob.glob(f'{INPUT_DIR}*.xml')):

    xml_doc = remove_namespaces(etree.parse(file))
    references = extract_references_from_document(xml_doc)
    
    if len(references) == 0:
        continue

    results.append(references)

INPUT_DIR = 'results/xml_sources/local_government_acts/'
for file in tqdm(glob.glob(f'{INPUT_DIR}*.xml')):

    xml_doc = remove_namespaces(etree.parse(file))
    references = extract_references_from_document(xml_doc)
    
    if len(references) == 0:
        continue

    results.append(references)

INPUT_DIR = 'results/xml_sources/government_regulations/'
for file in tqdm(glob.glob(f'{INPUT_DIR}*.xml')):

    xml_doc = remove_namespaces(etree.parse(file))
    references = extract_references_from_document(xml_doc)
    
    if len(references) == 0:
        continue

    results.append(references)

INPUT_DIR = 'results/xml_sources/government_orders/'
for file in tqdm(glob.glob(f'{INPUT_DIR}*.xml')):

    xml_doc = remove_namespaces(etree.parse(file))
    references = extract_references_from_document(xml_doc)
    
    if len(references) == 0:
        continue

    results.append(references)

external_links = concat(results, axis=0).drop_duplicates()
external_links.to_csv('results/external_links.csv', index=False)
